fix タスク更新処理系をまとめた
